Rename workflow.procfuncs to workflow.processing
