+++ bin/desi_resubmit_queue_failures
@@ -13,7 +13,7 @@ import glob
 ## Import some helper functions, you can see their definitions by uncomenting the bash shell command
 from desispec.workflow.tableio import load_table, write_table
 from desispec.workflow.proctable import get_processing_table_pathname
-from desispec.workflow.procfuncs import update_and_recurvsively_submit
+from desispec.workflow.processing import update_and_recurvsively_submit
 from desispec.workflow.queue import get_resubmission_states
 
 def parse_args():  # options=None):
